extract register_custom_staging_rule
