+++ doc/releases/changelog-dev.md
@@ -430,6 +430,9 @@
   of arrays with dynamic shapes.
   [(#7052)](https://github.com/PennyLaneAI/pennylane/pull/7052)
 
+* Adds a `qml.capture.register_custom_staging_rule` for handling higher order primitives
+  that return new dynamicly shaped arrays.
+
 <h3>Labs: a place for unified and rapid prototyping of research software 🧪</h3>
 
 * ``pennylane.labs.dla.lie_closure_dense`` is removed and integrated into ``qml.lie_closure`` using the new ``dense`` keyword.
